+++ app/frontend/components/CardsList/OptionCheckbox/index.js
