+++ app/frontend/components/CardsList/SortSelect/index.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react"
-import SelectField          from "../../../commons/SelectField"
+import SelectField          from "../../commons/SelectField"
 import * as utils           from "../cardsListUtils"
 
 export default class SortSelect extends Component {
