+++ app/frontend/components/CardsList/cardsListUtils.js
