+++ app/frontend/styles/partials/search.sass
@@ -24,7 +24,7 @@
   .result
     margin-top: 1.2rem
 
-@media screen and (min-width: $tablet)
+@media screen and (min-width: $desktop)
   .search
     .scores, .users
       display: flex
