+++ app/frontend/styles/partials/user-card.sass
@@ -30,11 +30,18 @@
         .profile
           margin-bottom: 1em
           p
-            font-size: .95em
+            font-size: .85em
             margin-bottom: 0
-        .scores-count
-          .icon
-            margin-right: .4em
+      nav
+        align-items: center
+        .control
+          font-size: .9rem
+          margin-top: .2em
+          &:not(:last-child)
+            margin-right: 1.2rem
+        .icon
+          margin-top: .1em
+          margin-right: .2em
     &:hover, &:active
       .box
         background-color: #fafafa
